Changed the updateDepartmentsCDIBean so that it gets the employee list from the database and not the department object since the department object was not refreshed to include new employees if these were added in the existing session.
Now assessments can be completed by the immediate manager. They also have a dedline and the conclusion is stored in the database. When the immediate manager completes the assessment then it is no longere available for updates.

Corrected bugs in the web page createEdrPage1.xhtml and the associated CDI bean. Also changed the employee converter in the same page to employeeConverter1 which uses first name, last name and email to identify and employee.