First version of the applicant interface
Applicants can add study and professional experience records.

Additionally minor fixes to the authorization process so that roles are assigned correctly
when a user does not login from the top right login menu but is presented with the login page

Finally
A CountriesList object to be used in various places where countries fields are available
Countries are presented in a drop down menu with the country code as value and the localized country name
The country code is stored in the db instead of the country name.
A DateRangeFutureValidator that checks that the end date is not in the future and that it is after the start date